+++ src/android/Impl.java
@@ -24,8 +24,11 @@ import android.accounts.AccountManager;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.ActivityInfo;
 import android.content.pm.PackageManager;
+import android.content.pm.ResolveInfo;
 import android.net.Uri;
+import android.os.Parcelable;
 import android.text.Html;
 import android.util.Log;
 import android.util.Patterns;
@@ -34,9 +37,11 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.regex.Pattern;
 
 import static android.content.Intent.ACTION_SENDTO;
+import static android.content.Intent.EXTRA_INITIAL_INTENTS;
 import static android.content.Intent.FLAG_ACTIVITY_CLEAR_TASK;
 import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
 import static android.content.Intent.FLAG_ACTIVITY_PREVIOUS_IS_TOP;
@@ -67,35 +72,56 @@ class Impl {
      * @return          The resulting intent.
      */
     Intent getDraft (JSONObject params) {
-        Intent mail = getEmailIntent();
-        String app  = params.optString("app", MAILTO_SCHEME);
-
-        if (params.has("subject"))
-            setSubject(params, mail);
-
-        if (params.has("body"))
-            setBody(params, mail);
-
-        if (params.has("to"))
-            setRecipients(params, mail);
-
-        if (params.has("cc"))
-            setCcRecipients(params, mail);
-
-        if (params.has("bcc"))
-            setBccRecipients(params, mail);
-
-        if (params.has("attachments"))
-            setAttachments(params, mail);
-
-        if (params.has("type"))
-            setType(params, mail);
+        Intent draft  = getFilledEmailIntent(params);
+        String app    = params.optString("app", MAILTO_SCHEME);
+        String header = params.optString("chooserHeader", "Open with");
 
         if (!app.equals(MAILTO_SCHEME) && isAppInstalled(app)) {
-            mail.setPackage(app);
+            return draft.setPackage(app);
         }
 
-        return mail;
+        List<Intent> targets = new ArrayList<>();
+
+        for (String clientId : getEmailClientIds()) {
+            Intent target = (Intent) draft.clone();
+            targets.add(target.setPackage(clientId));
+        }
+
+        return Intent.createChooser(targets.remove(0), header)
+                .putExtra(EXTRA_INITIAL_INTENTS, targets.toArray(new Parcelable[0]));
+    }
+
+    /**
+     * The intent with the containing email properties.
+     *
+     * @param params    The email properties like subject or body.
+     * @return          The resulting intent.
+     */
+    private Intent getFilledEmailIntent (JSONObject params) {
+        Intent draft = getEmailIntent();
+
+        if (params.has("subject"))
+            setSubject(params, draft);
+
+        if (params.has("body"))
+            setBody(params, draft);
+
+        if (params.has("to"))
+            setRecipients(params, draft);
+
+        if (params.has("cc"))
+            setCcRecipients(params, draft);
+
+        if (params.has("bcc"))
+            setBccRecipients(params, draft);
+
+        if (params.has("attachments"))
+            setAttachments(params, draft);
+
+        if (params.has("type"))
+            setType(params, draft);
+
+        return draft;
     }
 
     /**
@@ -238,6 +264,37 @@ class Impl {
         }
 
         return false;
+    }
+
+    /**
+     * Get the info for all available email client activities.
+     */
+    private List<ActivityInfo> getEmailClients() {
+        Intent           intent = getEmailIntent();
+        PackageManager       pm = ctx.getPackageManager();
+        List<ResolveInfo>  apps = pm.queryIntentActivities(intent, 0);
+        List<ActivityInfo> list = new ArrayList<>();
+
+        for (ResolveInfo app : apps) {
+            if (app.activityInfo.isEnabled()) {
+                list.add(app.activityInfo);
+            }
+        }
+
+        return list;
+    }
+
+    /**
+     * Get package IDs for all available email clients.
+     */
+    List<String> getEmailClientIds() {
+        List<String> ids = new ArrayList<>();
+
+        for (ActivityInfo app : getEmailClients()) {
+            ids.add(app.packageName);
+        }
+
+        return ids;
     }
 
     /**